[parser] brainfuck as a parser example
This is a turing-complete language running in the turing complete
constexpr context of a turing-complete language.

It effectively tests if the parser can be recursive and do some really
complex stuff. The implementation is too naive and brainfuck quickly
reaches the constexpr execution limit. It is a nice stress test for the
parser, but it is far away from a real implementation.

Nonetheless, it is super-cool to use an embedded turing-complete
language to define a C++ constant! :)