[parser] many/at_least_one with left priority
This introduces a new `<<` operator to allow `many` and `at_least_one`
to be defined with left priority. So, giving a parser like this:

```
    many(my_repeat(), my_init {})
```

That was the equivalent of: `my_init + (my_repeat + ...)` instead of
the more intuitive `my_init + my_repeat + ...`.

The new operator allows a more powerful composition like this:

```
    my_init() << my_repeat()
```

Which will use the result of `my_init` and aggregate it incrementally
with the result of `my_repeat`.

Quite proud of this! It is not a novel concept, but I manage to create
it by myself... :)